requirements.txt
pandas
numpy
plotly
streamlit
watchdog
python-dotenv
pyarrow
